improve: boss cooldown timer (#963)
Creating a new function to use in Lua to update Boss Cooldown in the Client.

If you use the mechanism created for the Lever() class, it takes care of this process automatically.

Example:
When killing the boss Duke Krule automatically, the Cooldown timer will be updated. But if you kill Boss Grand Master Oberon, he won't upgrade. This happens because the Oberon Lever update is done manually

Note:
For manual cases like above example. Oberon it is necessary to use the player:sendBosstiaryCooldownTimer() function after doing setStorageValue in the key timer of the boss.