fix: monster runAwayHealth respects multiplier (#1208)
Prior to this change, if you change the HP multiplier creatures like rabbit, deer, etc would follow players. This is incorrect behaviour since they are meant to run away even in full health.

This happens due to the absolute value based configuration for run away health, which doesn't play well with the multiplier. So this fix also multiples the runaway health by the same factor.