feat: loyalty system (#1213)
This commit implements the Loyalty System based on the code from pull request #425 (Thanks @marcosvf132). The Loyalty System functionality is inspired by the loyalty system in Tibia (https://tibia.fandom.com/wiki/Loyalty_System).

Enhances the loyalty system with new configuration options, accurate tracking of premium days, and support for customizable bonuses. The Loyalty System introduces several new configuration keys that can be customized:

• loyaltyEnabled: Determines whether the loyalty system is enabled or not (default: true).
• loyaltyPointsPerCreationDay: Sets the number of loyalty points awarded per day of account creation (default: 1).
• loyaltyPointsPerPremiumDaySpent: Sets the number of loyalty points awarded per premium day spent (default: 0).
• loyaltyPointsPerPremiumDayPurchased: Sets the number of loyalty points awarded per premium day purchased (default: 0).
• loyaltyBonusPercentageMultiplier: Allows for customization of loyalty bonus strength (default: 1.0).

By default, the loyaltyPointsPerCreationDay is set to 1, providing 1 loyalty point per day since the account was created. This conservative default setting ensures a slow accumulation of loyalty points, similar to the original Tibia loyalty system. Server administrators can adjust these settings as needed.

To accurately track premium days spent, a new column was added to the accounts table, which stores the total number of premium days purchased on the account.

The implementation includes support for awarding loyalty points at the moment of VIP purchase. If a server uses a different mechanism for VIP, they can adapt the loyalty system accordingly.

The loyaltyBonusPercentageMultiplier allows administrators to adjust the strength of bonuses. The default settings are relatively mild, so this option provides flexibility. For example, to award a 10 skill bonus at around magic level or skill 120, the multiplier should be set to 4.0.

To ensure accurate calculations, a new library 'absl' was introduced. This enables precise calculation of bonuses as loyalty points are multiplied by the bonus percentage. This is necessary because only the current level's "tries" (or mana spent) are tracked, and calculations may exceed the limits of uint64_t.