feat: boss specific multipliers (#1266)
This is both a bugfix and a new feature. Found this out when testing Urmahlulu with a health multiplier > 1.0 ; its stages are janky as heck. This fixes that, but also splits up the boss health multiplier into its own, so admins can configure mobs and bosses separately.

Note: this PR also deletes some unused code left from the old hazard system that is no longer active. I didn't create a separate PR for that since it was already on this commit, but I can split it up if reviewers prefer.