feat: introduce db transaction guard for muto-managed transactions (#738)
Introduces the DBTransactionGuard class, which provides automatic management of a DBTransaction object. This class is designed to commit the transaction automatically when its instance goes out of scope, unless a rollback is explicitly requested. 

The DBTransactionGuard is responsible for handling exceptions during the commit process, logging them as errors. To ensure proper usage, this class has been made non-copyable, thus enforcing that it can only be tied to a single transaction at a time.