fix: show correct fluid names (#1355)
When we moved to loading appearances.dat we started ignoring the fluids from items.xml. This was done as a protection when people add new items to items.xml that aren't in appearances.dat, but it incorrectly filtered out fluids.

Also: it was previously possible to sell the vial/bucket/whatever on the market because fluid ID was some dirty memory space, that is now no longer possible since you can no longer obtain the broken fluid.