improve: Move all logging usages to g_logger
SPDLOG is just one logging library. While it's very handy and we use it as our default logging solution
we might want to use a different solution in the future. That said, using g_logger linked to a DI injection
give us the flexibility to replace the underlying implementation in the future.

Also, the current macros provided by SPDLOG do not rely in the standard log level, but in a self managed
compiler time definition for log level, that is not customisable.