feat: Move CMake sources to subdirectories
CMake is designed to be modular, meaning that each subdirectory is
aware and responsible of it own needs, dependencies and includes.

Previously, it was really hard to make this separation because our
CMake files were huge and organized in a kinda of all-in-one file.

With the recent CMake improvements, we can leverage CMake sub directories
to make our includes more moduler and easier to read, instead of
having a one imense single source of truth file.