feat: implement docker-compose (#1252)
Implement docker-compose to streamline the steps to launch a server. With just one command, you will have a server ready. Add more options to the existing scripts to adapt them for both Docker and for those who have configured VCPKG differently from the tutorial.

Keep the old Dockerfile, as it serves a different purpose.