fix: Guarantees that canary_lib is always static (#1483)
# Description

There was a small oversight in the CMake improvements, we've made the
CanaryLib configurable between shared and static, and right now we are
not ready for this change, it needs to be always static. One can still
use `BUILD_STATIC_LIBRARY` to choose if the external dependencies should
be linked statically or dynamically, but the generated CanaryLib will
always be static.