fix: house transfer owner only in startup (#1621)
The transfer of houses is now permitted only during server restarts to
prevent server freezing when moving items to the depot. While this is
not the most sophisticated solution, it's the simplest and least
error-prone due to the base's complexity. This approach mirrors what
CipSoft employs in the global version of Tibia and effectively addresses
the issue without introducing possible new bugs.

Additional updates include modifications to NPCs to sell the new kit,
which displays the item's sprite in the wrap container. Support for
using old kits remains, allowing players to convert them into new ones.

Unrelated changes affecting houses:
• Purchasing a house with items still inside is no longer possible. This
prevents issues in rare scenarios, such as incomplete item transfers due
to console closure.
• Introduced an "all" parameter to the "/r" command for removing all
items from a tile.
• Removed the "allowpickUpAble" tags from ground/border and derived
items. These were initially implemented to enable item "throwing" onto
tiles. Now, an item can be thrown only if its type is "trashholder,"
fulfilling the same purpose. This change was made because the check for
items inside the house before purchasing relies on the "pickupable" tag,
which is not applicable to tiles and derivatives.