fix: stack overflow in bug report (#1704)
fix the FS.mkdir_p(path) function that was causing stackoverflow
exception

I made a change so that the code works as follows: The path is divided
into parts, then the function goes through the components, creating the
directory if it does not exist. The process repeats until all components
of the path have been successfully completed, avoiding stackoverflow
exception. At the end, the function returns true to indicate that the
operation was successful.