improve: normal random skew towards means (#1672)
The current code for normal distribution is skewed and tends to generate much more times the central element than the rest of the distribution, more than is expected of a normal distribution.

The proper solution is to just discard the result and generate a new value. The while loop will only run once in 95% of the calls. There is an infinitely small chance of an infinite loop.

The new code is also roughly 30% faster (on a Ryzen 7 4800H chip) due to it performing less branching and having only one return.