fix: disable assert for not debug compilaton and add logs (#2182)
• Added the NDEBUG macro to the processor definitions to disable assert
in non-debug compilations.
• Implemented logging and return statements to handle errors instead of
using assert. This is necessary to prevent abrupt interruptions in
production servers.
• Made several corrections and adjustments in the build solution,
including the use of AddressSanitizer (ASan) in debug builds.

Resolves #2156

Using assert is not advisable in production because when an assertion
fails, the program is immediately terminated, causing uncontrolled
disruptions on the server. Instead, it's preferable to use proper error
handling mechanisms, such as logging and exceptions, to deal with issues
in a controlled manner and maintain system stability in production.