improve: async batch dispatcher (#2666)
Now the asynchronous requests will be made in batches and not one at a
time, that is, we will take the total number of tasks divided by the
number of available threads, this avoids locking the threads to retrieve
the next task to be processed.

ex: if we have 1000 tasks and 8 threads, 125 tasks will be triggered per
thread. (1000/8) = 125

In this PR I also brought a new method called asyncWait, with this
method you can make asynchronous requests but without needing to define
a group, the asynchronous tasks will be executed immediately and will
wait for them to finish.