perf: improve npc interactions (#3021)
Improves the npc interactions following the Global Tibia, specifically
the npc turning to the last talked player when him moves and turning to
the next player that last talked to him. It improves as well the
checkage of the interaction with players that was doing on every npc
movement/tick (on think), now it checks only on actions like, player
entering in a new topic, greeting, ungreeting, etc.
Added a empty checkage in the map before doing the find to avoid
unnecessary operations done on find.