fix: resolve nil value errors in handleGuildWar function (#3172)
I made corrections to the `death.lua` script to resolve nil value errors
in the `handleGuildWar` function. I added checks to ensure that both the
player and the killer are valid and that both belong to a guild before
proceeding with the guild war logic. This helps prevent failures when
one of the objects is undefined.

Fixes from: https://github.com/opentibiabr/canary/commit/b07e5a7272144dc09fb471ce5368dec323e08319